requirements.txt
aiosqlite
discord.py
tqdm
typeguard